added react-query and axios for fetching
