+++ package.json
@@ -14,7 +14,8 @@
     "@types/powerapps-component-framework": "^1.2.3",
     "@types/react": "^16.9.35",
     "@uifabric/react-hooks": "^7.4.4",
-    "use-http": "^1.0.12"
+    "axios": "^0.21.1",
+    "react-query": "^3.12.2"
   },
   "devDependencies": {
     "pcf-scripts": "^1",
